requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6

langchain==0.2.14
langchain-core==0.2.32
langchain-community==0.2.12
langchain-astradb==0.4.0
langchain-groq==0.1.6

python-dotenv==1.0.1